agents/deployment_agent/requirements.txt
redis>=4.5.1
openai>=0.27.0
requests>=2.28.0
python-dotenv>=1.0.0
crewai>=0.28.0
langchain-openai>=0.0.2
kubernetes>=26.1.0
PyYAML>=6.0
GitPython>=3.1.30
# ArgoCD specific dependencies
argocd-python-client>=0.0.1
urllib3>=1.26.0
langchain==0.0.267
openai==0.27.8
python-dotenv==1.0.0
requests==2.31.0
GitPython==3.1.32
PyYAML==6.0
crewai==0.28.5
langchain-openai==0.0.2
nats-py==2.4.0